(refactor): refactored network tracing using a network.rs module
