+++ core/src/components/metrics_tracer/src/mod.rs
@@ -2,3 +2,4 @@ mod bindings;
 mod cpu;
 mod data_structures;
 mod memory;
+mod network;
